requirements.txt
Package                 Version
----------------------- -------------------
absl-py                 0.13.0
argon2-cffi             21.1.0
astor                   0.8.1
astunparse              1.6.3
async-generator         1.10
attrs                   21.2.0
autokeras               1.0.16
backcall                0.2.0
bleach                  4.1.0
cached-property         1.5.2
cachetools              4.2.2
certifi                 2021.5.30
cffi                    1.14.6
charset-normalizer      2.0.6
clang                   5.0
colorama                0.4.4
cycler                  0.10.0
dataclasses             0.8
decorator               5.1.0
defusedxml              0.7.1
entrypoints             0.3
flatbuffers             1.12
gast                    0.4.0
glob2                   0.7
google-auth             1.35.0
google-auth-oauthlib    0.4.6
google-pasta            0.2.0
grpcio                  1.34.1
h5py                    3.1.0
idna                    3.2
image-slicer            2.1.1
importlib-metadata      4.8.1
ipykernel               5.5.5
ipython                 7.16.1
ipython-genutils        0.2.0
ipywidgets              7.6.5
jedi                    0.18.0
Jinja2                  3.0.1
joblib                  1.0.1
jsonschema              3.2.0
jupyter                 1.0.0
jupyter-client          7.0.2
jupyter-console         6.4.0
jupyter-core            4.7.1
jupyterlab-pygments     0.1.2
jupyterlab-widgets      1.0.2
keras                   2.6.0
Keras-Applications      1.0.8
keras-nightly           2.5.0.dev2021032900
Keras-Preprocessing     1.1.2
keras-tuner             1.0.4
kiwisolver              1.3.1
kt-legacy               1.0.4
Markdown                3.3.4
MarkupSafe              2.0.1
matplotlib              3.3.4
mistune                 0.8.4
mlxtend                 0.19.0
nbclient                0.5.4
nbconvert               6.0.7
nbformat                5.1.3
nest-asyncio            1.5.1
notebook                6.4.4
numpy                   1.19.5
oauthlib                3.1.1
opencv-python           4.5.3.56
opt-einsum              3.3.0
packaging               21.0
pandas                  1.1.5
pandocfilters           1.5.0
parso                   0.8.2
patsy                   0.5.1
pickleshare             0.7.5
Pillow                  7.2.0
pip                     21.0.1
prettytable             2.2.1
prometheus-client       0.11.0
prompt-toolkit          3.0.20
protobuf                3.18.0
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycparser               2.20
pydot                   1.4.2
Pygments                2.10.0
pyparsing               2.4.7
pyrsistent              0.18.0
python-dateutil         2.8.2
pytz                    2021.1
pywin32                 301
pywinpty                1.1.4
pyzmq                   22.3.0
qtconsole               5.1.1
QtPy                    1.11.1
requests                2.26.0
requests-oauthlib       1.3.0
rsa                     4.7.2
scikit-learn            0.24.2
scipy                   1.2.0
seaborn                 0.11.2
Send2Trash              1.8.0
setuptools              52.0.0.post20210125
six                     1.15.0
statsmodels             0.12.2
tensorboard             2.6.0
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.0
tensorflow              2.5.0
tensorflow-estimator    2.5.0
termcolor               1.1.0
terminado               0.12.1
testpath                0.5.0
tf-slim                 1.1.0
threadpoolctl           2.2.0
tornado                 6.1
traitlets               4.3.3
typing-extensions       3.7.4.3
urllib3                 1.26.6
wcwidth                 0.2.5
webencodings            0.5.1
Werkzeug                2.0.1
wheel                   0.37.0
widgetsnbextension      3.5.1
wincertstore            0.2
wrapt                   1.12.1
xgboost                 1.4.2
zipp                    3.5.0
